+++ lib/persistentqueue/fastqueue_timing_test.go
@@ -2,8 +2,9 @@ package persistentqueue
 
 import (
 	"fmt"
-	"runtime"
 	"testing"
+
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/cgroup"
 )
 
 func BenchmarkFastQueueThroughputSerial(b *testing.B) {
@@ -36,7 +37,7 @@ func BenchmarkFastQueueThroughputConcurrent(b *testing.B) {
 			b.SetBytes(int64(blockSize) * iterationsCount)
 			path := fmt.Sprintf("bench-fast-queue-throughput-concurrent-%d", blockSize)
 			mustDeleteDir(path)
-			fq := MustOpenFastQueue(path, "foobar", iterationsCount*runtime.GOMAXPROCS(-1)*2, 0)
+			fq := MustOpenFastQueue(path, "foobar", iterationsCount*cgroup.AvailableCPUs()*2, 0)
 			defer func() {
 				fq.MustClose()
 				mustDeleteDir(path)