+++ lib/protoparser/influx/parser_test.go
@@ -468,6 +468,36 @@ func TestRowsUnmarshalSuccess(t *testing.T) {
 		},
 	})
 
+	// Superfluous whitespace between tags, fields and timestamps.
+	f(`cpu_utilization,host=mnsbook-pro.local value=119.8 1607222595591`, &Rows{
+		Rows: []Row{{
+			Measurement: "cpu_utilization",
+			Tags: []Tag{{
+				Key:   "host",
+				Value: "mnsbook-pro.local",
+			}},
+			Fields: []Field{{
+				Key:   "value",
+				Value: 119.8,
+			}},
+			Timestamp: 1607222595591,
+		}},
+	})
+	f(`cpu_utilization,host=mnsbook-pro.local   value=119.8   1607222595591`, &Rows{
+		Rows: []Row{{
+			Measurement: "cpu_utilization",
+			Tags: []Tag{{
+				Key:   "host",
+				Value: "mnsbook-pro.local",
+			}},
+			Fields: []Field{{
+				Key:   "value",
+				Value: 119.8,
+			}},
+			Timestamp: 1607222595591,
+		}},
+	})
+
 	f("x,y=z,g=p:\\ \\ 5432\\,\\ gp\\ mon\\ [lol]\\ con10\\ cmd5\\ SELECT f=1", &Rows{
 		Rows: []Row{{
 			Measurement: "x",